chore(fast-element): cleanup some subscriber interfaces and organization
