+++ packages/web-components/fast-element/src/observation/array-observer.ts
@@ -1,7 +1,6 @@
 import { DOM } from "../dom";
 import { Observable } from "./observable";
-import { Subscriber, SubscriberCollection } from "./subscriber-collection";
-import { Notifier } from "./notifier";
+import { SubscriberCollection } from "./notifier";
 import {
     calcSplices,
     newSplice,
@@ -30,14 +29,12 @@ function adjustIndex(changeRecord: Splice, array: any[]): Splice {
     return changeRecord;
 }
 
-export class ArrayObserver extends SubscriberCollection implements Notifier {
+export class ArrayObserver extends SubscriberCollection {
     private collection: any[];
     private oldCollection: any[] | undefined = void 0;
     private splices: any[] | undefined = void 0;
     private needsQueue: boolean = true;
-
-    subscribe: (subscriber: Subscriber) => void = this.addSubscriber;
-    unsubscribe: (subscriber: Subscriber) => void = this.removeSubscriber;
+    call: () => void = this.flush;
 
     constructor(collection: any[]) {
         super();
@@ -45,7 +42,7 @@ export class ArrayObserver extends SubscriberCollection implements Notifier {
         this.collection = collection;
     }
 
-    addSplice(splice: Splice): void {
+    public addSplice(splice: Splice): void {
         if (this.splices === void 0) {
             this.splices = [splice];
         } else {
@@ -58,7 +55,7 @@ export class ArrayObserver extends SubscriberCollection implements Notifier {
         }
     }
 
-    reset(oldCollection: any[] | undefined): void {
+    public reset(oldCollection: any[] | undefined): void {
         this.oldCollection = oldCollection;
 
         if (this.needsQueue) {
@@ -67,15 +64,13 @@ export class ArrayObserver extends SubscriberCollection implements Notifier {
         }
     }
 
-    notify(): void {
-        if (this.splices !== void 0 || this.oldCollection !== void 0) {
-            this.call();
-        }
-    }
-
-    call(): void {
+    public flush(): void {
         const splices = this.splices;
         const oldCollection = this.oldCollection;
+
+        if (splices === void 0 && oldCollection === void 0) {
+            return;
+        }
 
         this.needsQueue = true;
         this.splices = void 0;
@@ -93,7 +88,7 @@ export class ArrayObserver extends SubscriberCollection implements Notifier {
                       oldCollection.length
                   );
 
-        this.notifySubscribers(this, finalSplices);
+        this.notify(this, finalSplices);
     }
 }
 
@@ -152,7 +147,7 @@ export function enableArrayObservation(): void {
         const o = (this as any).$fastController as ArrayObserver;
 
         if (o !== void 0) {
-            o.notify();
+            o.flush();
             oldArray = this.slice();
         }
 
@@ -182,7 +177,7 @@ export function enableArrayObservation(): void {
         const o = (this as any).$fastController as ArrayObserver;
 
         if (o !== void 0) {
-            o.notify();
+            o.flush();
             oldArray = this.slice();
         }
